[librptexture] rp_image: Change palette_len from int to unsigned int.
Code size difference: (64-bit Gentoo Linux, gcc-11.2.0, release build, no LTO)

   text    data     bss     dec     hex filename
1496808   93624   30056 1620488  18ba08 bin/rpcli
1496808   93624   30056 1620488  18ba08 bin/rpcli
      0       0       0       0       0 Difference

   text    data     bss     dec     hex filename
1690664   99000   30024 1819688  1bc428 lib/rom-properties-gtk3.so
1690600   99000   30024 1819624  1bc3e8 lib/rom-properties-gtk3.so
    -64       0       0     -64     -40 Difference

   text    data     bss     dec     hex filename
1681506   98600   30024 1810130  1b9ed2 lib/rom-properties-gtk4.so
1681458   98600   30024 1810082  1b9ea2 lib/rom-properties-gtk4.so
    -48       0       0     -48     -30 Difference

   text    data     bss     dec     hex filename
1684273   98848   29992 1813113  1baa79 lib/rom-properties-xfce.so
1684209   98848   29992 1813049  1baa39 lib/rom-properties-xfce.so
    -64       0       0     -64     -40 Difference

   text    data     bss     dec     hex filename
1987186  113424   29608 2130218  20812a lib/rom-properties-kf5.so
1987138  113424   29608 2130170  2080fa lib/rom-properties-kf5.so
    -48       0       0     -48     -30 Difference

+++ src/kde/RpQImageBackend.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (KDE4/KF5)                         *
  * RpQImageBackend.cpp: rp_image_backend using QImage.                     *
  *                                                                         *
- * Copyright (c) 2016-2020 by David Korth.                                 *
+ * Copyright (c) 2016-2021 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -10,9 +10,9 @@
 #include "RpQImageBackend.hpp"
 
 #if QT_VERSION >= QT_VERSION_CHECK(4,0,0) && QT_VERSION < QT_VERSION_CHECK(5,0,0)
-# include "QImageData_qt4.hpp"
+#  include "QImageData_qt4.hpp"
 #elif QT_VERSION < QT_VERSION_CHECK(4,0,0)
-# error Unsupported Qt version.
+#  error Unsupported Qt version.
 #endif
 
 // librpbase, librptexture
@@ -140,9 +140,9 @@ const uint32_t *RpQImageBackend::palette(void) const
 	return m_qPalette.constData();
 }
 
-int RpQImageBackend::palette_len(void) const
+unsigned int RpQImageBackend::palette_len(void) const
 {
-	return m_qPalette.size();
+	return static_cast<unsigned int>(m_qPalette.size());
 }
 
 /**
